+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Unreleased changes
+* bugfix: `earthmover deps` failed to find nested local packages
+
 ### v0.4.0
 <details>
 <summary>Released 2024-10-16</summary>